backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
scipy
ortools
dimod
dwave-system
pytest
